lambda/requirements.txt
slack-sdk==3.20.2
PyYAML==5.4.1
pytz==2022.1